requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
